+++ src/pjpl/s7/common/ConstProcess.java
@@ -5,6 +5,6 @@ package pjpl.s7.common;
  * @author pjanczura
  */
 public class ConstProcess {
-	public final static int PROCESS1_ID = 1;
+	public final static byte PROCESS1_ID = 1;
 	public final static String PROCESS1_NAME = "Process 1";
 }
